requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
python-dotenv==1.0.0
magic-filter==1.0.9
multidict==6.0.4
pycountry==22.3.5
pytz==2023.3
requests==2.31.0
six==1.16.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
yarl==1.9.2